[REFACTOR, TEST] Order, OrderMenu 리팩토링 및 테스트
