+++ src/main/java/bleuauction/bleuauction_be/server/order/controller/OrderController.java
@@ -1,6 +1,5 @@
 package bleuauction.bleuauction_be.server.order.controller;
 
-import bleuauction.bleuauction_be.server.member.entity.Member;
 import bleuauction.bleuauction_be.server.member.service.MemberService;
 import bleuauction.bleuauction_be.server.order.entity.Order;
 import bleuauction.bleuauction_be.server.order.repository.OrderRepository;
@@ -8,16 +7,12 @@ import bleuauction.bleuauction_be.server.order.service.OrderService;
 import bleuauction.bleuauction_be.server.util.CreateJwt;
 import bleuauction.bleuauction_be.server.util.TokenMember;
 import jakarta.servlet.http.HttpSession;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
-
-import java.util.List;
+import java.util.Optional;
 
 @Slf4j
 @RestController
@@ -39,17 +34,16 @@ public class OrderController {
 
 
   @PostMapping("/new")
-  public ResponseEntity<String> order(Order order, HttpSession session) {
-    orderService.enroll(order);
-    session.setAttribute("order", order);
+  public ResponseEntity<?> addorder(Order order, HttpSession session) throws Exception{
     log.info("order/postnew");
-    return ResponseEntity.status(HttpStatus.CREATED).body("Order created successfully");
+    session.setAttribute("order", order);
+    return orderService.addOrder(order);
   }
 
 
   //회원별 주문 조회
   @GetMapping
-  public ResponseEntity<?> findOrders(@RequestHeader("Authorization") String  authorizationHeader) {
+  public ResponseEntity<?> findOrders(@RequestHeader("Authorization") String authorizationHeader) {
 
     ResponseEntity<?> verificationResult = createJwt.verifyAccessToken(authorizationHeader, createJwt);
     if (verificationResult != null) {
@@ -58,74 +52,41 @@ public class OrderController {
 
     TokenMember tokenMember = createJwt.getTokenMember(authorizationHeader);
     log.info("token: " + tokenMember);
-    Optional<Member> loginUser = memberService.findByMemberNo(tokenMember.getMemberNo());
 
-    if (loginUser == null) {
-      return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인한 사용자가 아닙니다.");
-    }
-
-    List<Order> orders = orderRepository.findByOrderMenusMemberMemberNo(loginUser.get());
-
-    if (orders.isEmpty()) {
-      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("주문이 없습니다.");
-    } else {
-      for (Order order : orders) {
-        order.calculateOrderPrice(); // 주문 가격 계산
-      }
-      return ResponseEntity.ok(orders);
-    }
+    return orderService.findOrdersByMemberNo(tokenMember.getMemberNo());
   }
 
   // 가게(가게주인)별 주문 조회
   @GetMapping("/store")
-  public ResponseEntity<?> findOrdersbyStore( @RequestHeader("Authorization") String  authorizationHeader,HttpSession session) {
+  public ResponseEntity<?> findOrdersbyStore(@RequestHeader("Authorization") String  authorizationHeader,HttpSession session) {
     ResponseEntity<?> verificationResult = createJwt.verifyAccessToken(authorizationHeader, createJwt);
     if (verificationResult != null) {
       return verificationResult;
     }
 
     TokenMember tokenMember = createJwt.getTokenMember(authorizationHeader);
-    Optional<Member> loginUser = memberService.findByMemberNo(tokenMember.getMemberNo());
+    log.info("token: " + tokenMember);
 
-    Long memberNo = loginUser.get().getMemberNo();
-
-
-    if (loginUser.get() == null) {
-      return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인한 사용자가 아닙니다.");
-    }
-
-    List<Order> orders = orderRepository.findOrdersByMemberAndStore(memberNo);
-
-    if (orders.isEmpty()) {
-      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("주문이 없습니다.");
-    } else {
-      return ResponseEntity.ok(orders);
-    }
+    return orderService.findOrdersByMemberAndStore(tokenMember.getMemberNo());
   }
 
   // 삭제--오더메뉴도 같이
-  @PutMapping("/delete/{orderNo}")
+  @PostMapping("/delete/{orderNo}")
   public ResponseEntity<String> deleteOrder(@PathVariable("orderNo") Long orderNo) {
-    Order order = orderService.findOne(orderNo);
-    if (order != null) {
-      order.delete();
-      orderService.enroll(order);
-    }
-    return ResponseEntity.ok("Order deleted successfully");
+      return orderService.deleteOrder(orderNo);
   }
 
   //디테일(수정)
   @GetMapping("detail/{orderNo}")
-  public ResponseEntity<Order> detailOrder(HttpSession session, @PathVariable("orderNo") Long orderNo) {
-    Order order = orderService.findOne(orderNo);
+  public ResponseEntity<?> detailOrder(HttpSession session, @PathVariable("orderNo") Long orderNo) {
+    Optional<Order> order = orderService.findOne(orderNo);
     session.setAttribute("order", order);
     return ResponseEntity.ok(order);
   }
 
   @PutMapping("/update/{orderNo}")
-  public ResponseEntity<String> updateOrder (Order order, @PathVariable("orderNo") Long orderNo) {
-    orderService.update(order);
+  public ResponseEntity<String> updateOrder (@PathVariable("orderNo") Long orderNo) {
     log.info("order/update");
-    return ResponseEntity.ok("Order updated successfully");
+    return orderService.update(orderNo);
   }
 }
